Pallindrome Index
Problem Statement

You are given a string of lower case letters. Your task is to figure
out the index of the character on whose removal it will make the string
a palindrome. There will always be a valid solution.

In case the string is already a palindrome, then -1 is also a valid
answer along with possible indices.